refactor: 파일 구조 변경, 접근 제어자 변경 #6
- Double extension을 BankPrinter로 이동하여 private extension으로 사용(특정 기능 제공)
- 위임할 기능을 나타내는 protocol을 각 타입의 파일에 위치하도록 변경
